% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annoy_knn.R
\name{annoy-knn}
\alias{annoy-knn}
\alias{annoy_build}
\alias{annoy_search}
\title{Helper functions for RcppAnnoy}
\usage{
annoy_build(
  data,
  metric = c("euclidean", "cosine", "manhattan", "hamming"),
  n.trees = 50,
  seed.use = 42,
  verbose = TRUE,
  by.row = TRUE
)

annoy_search(
  index,
  query,
  k,
  search.k = -1,
  include.distance = TRUE,
  n.threads = 1,
  chunk.size = 3000L,
  by.row = TRUE
)
}
\arguments{
\item{data}{A numeric matrix of data to build the 'annoy' index.}

\item{metric}{Type of distance calculation to use.}

\item{n.trees}{Builds a forest of \code{n.trees} trees. More trees gives higher
precision when querying.}

\item{seed.use}{Set a random seed. Setting \code{NULL} will not set a seed.}

\item{verbose}{Print progress.}

\item{by.row}{If \code{TRUE} (default), each row of input matrix is treated as an
observation. Otherwise, each observation is stored by column.}

\item{index}{An \code{.doc_links("AnnoyIndex")} output by \code{annoy_build}.}

\item{query}{A numeric matrix of data to search for neighbors.}

\item{k}{Number of neighbors to return. Must be within the size of each item
in \code{index}.}

\item{search.k}{During the query, it will inspect up to \code{search.k} nodes
which defaults to \code{n.trees} * \code{k} if not provided.}

\item{include.distance}{Whether or not to return the distances of kNN.}

\item{n.threads}{Number of threads to use.}

\item{chunk.size}{Used when \code{n.threads > 1}. How many query observations to
be processed in each chunk.}
}
\value{
\itemize{
\item \code{annoy_build}: A built \code{\link[RcppAnnoy]{AnnoyIndex}}.
}

\itemize{
\item \code{annoy_search}: A list containing \code{idx} matrix (and \code{dist} matrix if
\code{include.distance} is \code{TRUE})
}
}
\description{
Helper functions for building index of \pkg{RcppAnnoy} and search kNNs.
}
\examples{
data(pca)
ann <- annoy_build(pca)
nn <- annoy_search(ann, pca[1:20, ], k = 8)

}
\seealso{
\itemize{
\item \code{\link[RcppAnnoy]{AnnoyIndex}}
}
}
