% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{easy.NN-package}
\alias{easy.NN}
\alias{easy.NN-package}
\title{easy.NN: Manipulation for nearest neighbors (NNs)}
\description{
Some functions for data coercing, manipulation and calculation of nearest neighbors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ycli1995/easy.NN}
  \item Report bugs at \url{https://github.com/ycli1995/easy.NN/issues}
}

}
\author{
\strong{Maintainer}: Yuchen Li \email{ycli1995@outlook.com}

}
\keyword{internal}
