% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/knnToSNN.R
\name{knnToSNN}
\alias{knnToSNN}
\alias{knnToSNN.CsparseMatrix}
\alias{knnToSNN.list}
\title{Compute shared nearest neighbors from kNN results}
\usage{
knnToSNN(object, ...)

\method{knnToSNN}{CsparseMatrix}(object, prune = 1/15, repr = c("C", "T"), ...)

\method{knnToSNN}{list}(object, prune = 1/15, repr = c("C", "T"), ...)
}
\arguments{
\item{object}{An object storing kNN results.}

\item{...}{Arguments passed to other metheds.}

\item{prune}{Sets the cutoff for acceptable Jaccard index when computing the
neighborhood overlap for the SNN construction. Any edges with values less
than or equal to this will be set to 0 and removed from the SNN graph.
Essentially sets the stringency of pruning (0 — no pruning, 1 — prune
everything).}

\item{repr}{Return \code{CsparseMatrix} or \code{TsparseMatrix}.}
}
\value{
A sparse matrix representing the SNN graph.
}
\description{
Compute shared nearest neighbors from kNN results
}
\examples{
nn1 <- hnswNN(pca, k = 8)
nn1 <- nnToGraph(nn1, use.weights = FALSE)
snn <- knnToSNN(nn1)

nn2 <- hnswNN(pca, k = 8)
snn <- knnToSNN(nn2)  # Use the nn list directly

}
