% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{pca}
\alias{pca}
\title{A small example of PCA embeddings from single-cell RNA data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 80 rows and 19 columns.
}
\usage{
pca
}
\description{
A matrix of PCA embeddings. Each row represents a cell, and each column
represents a dimension in principal component  space. This matrix was
obtained from \code{pbmc_small} dataset in \pkg{SeuratObject}.
}
\keyword{datasets}
